Add new home page to start migration to new page
